Update AI.cs
Changed lowercase conversion method from .ToLower() to .ToLowerInvariant() in order to make it work with Turkish regional settings. If not forced to make an Invariant conversion, .NET converts capital "I" to lowercase "ı" which is linguistically correct for Turkish but fails for file paths thus leading the "can't load AI models for this implementation!" error.